+++ site/dat/docs/Changelog.md
@@ -5,6 +5,7 @@
  - selenium script from requests format
  - change file search algo to use paths relative to config locations
  - allow having script files relative to config locations
+ - allow having `included-configs` instruction
 
 ## 0.4.1
  - fix Locust.io search paths